backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
pandas>=2.0
reportlab>=4.0
